backend/requirements.txt
# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6
sse_starlette>=1.6.5

# Database
motor>=3.3.1  # MongoDB async driver
redis>=5.0.1
beanie>=1.23.0  # MongoDB ODM

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# LLM and Tools
openai>=1.0.0
langchain>=0.0.350
huggingface_hub>=0.19.0
chromadb>=0.4.15
tiktoken>=0.5.1
playwright>=1.39.0

# Multilingual Support
polyglot>=16.7.4
PyICU>=2.11
pycld2>=0.41

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.0
black>=23.10.0
isort>=5.12.0
mypy>=1.6.1

# Docker Support
docker>=6.1.3
python-docker>=6.1.3

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tenacity>=8.2.3
websockets>=11.0.3
